Added Patch revert Setting
Added a way to revert the old Patched file if ever needed. using settings.lua. Plus a way to notify if /when the patch was applied/reverted and needs a restart

though if the mod updated then this wont really matter as a resub will be needed of wand_DBG anyway